requirements.txt.txt
flask==2.2.2
flask-restful==0.3.9
mysql-connector-python==8.0.30